feat(product): client homepage layout
