+++ components/Header.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+  <header>
+    <h1 class="text-xl font-bold text-[#B7C0C2] mt-3 ml-3">La Corn'Eye</h1>
+  </header>
+</template>
+
+<style scoped>
+
+</style>
