+++ pages/index.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+import Home from "~/pages/Home.vue";
+</script>
+
+<template>
+  <Home />
+</template>
+
+<style scoped>
+
+</style>
