+++ plugins/vuetify.ts
@@ -1,0 +1,12 @@
+import { createVuetify } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+
+export default defineNuxtPlugin(nuxtApp => {
+    const vuetify = createVuetify({
+        components,
+        directives,
+    })
+
+    nuxtApp.vueApp.use(vuetify)
+})
